Split LuaUtils.cpp into topical files
